+++ GitHubNotificationManager/Components/Image/ImageLoaderView.swift
@@ -11,17 +11,16 @@ import GitHubNotificationManagerNetwork
 
 fileprivate typealias Container = VStack // FIXME: Group does not call onAppear
 struct ImageLoaderView : View {
-    @State var viewModel = ImageLoaderViewModel()
+    @ObservedObject var viewModel = ImageLoaderViewModel()
     
     let url: URLConvertible
     var body: some View {
         Container {
-            if self.viewModel.image != nil {
-                Image(uiImage: self.viewModel.image!)
-            } else {
-                EmptyView()
-            }
+            Image(uiImage: viewModel.image)
         }
+        .onReceive(viewModel.objectWillChange, perform: { (_) in
+            
+        })
         .onAppear {
             self.viewModel.load(url: self.url)
         }
