+++ GitHubNotificationManager/Frameworks/GitHubNotificationManagerNetwork/GitHub/GitHubAPI.swift
@@ -31,7 +31,7 @@ public struct GitHubAPI {
                 URLSession.shared.dataTaskPublisher(for: try request.urlRequest())
                     .map { $0.data }
                     .decode(type: Output.self, decoder: JSONDecoder())
-                    .print()
+//                    .print()
                     .receive(on: DispatchQueue.main)
                     .subscribe(subscriber)
             } catch {
